+++ src-cards/untangled/client/card_ui.cljs
@@ -2,8 +2,7 @@
   (:require
     [devtools.core :as devtools]
     untangled.client.fancy-defui
-    untangled.client.initial-app-state-card
-    untangled.client.intro))
+    untangled.client.initial-app-state-card))
 
 (devtools/enable-feature! :sanity-hints)
 (devtools/install!)