+++ src-cards/untangled/client/intro.cljs
@@ -1,8 +1,0 @@
-(ns untangled.client.intro
-  (:require
-    [devcards.core :as rc :refer-macros [defcard]]
-    [om.dom :as dom]))
-
-(defcard intro-card
-  "#Intro to Devcards!"
-  (dom/div nil "Hello from devcards & om!"))